requirements.txt
# write pip install -r requirements.txt in terminal to install all the libraries

# WebSocket server
websockets

# Async event loop support for nested environments (like Jupyter or async dev)
nest_asyncio

# JWT authentication
PyJWT

# Machine learning
scikit-learn
numpy
pandas

# PostgreSQL / TimescaleDB
psycopg2-binary

# Environment variable support
python-dotenv

# plots
matplotlib

# environments
dotenv